Making the functions in Julia
A goal at the start of the project was to get better at julia, cause I like julia.
Problem is that I did so much of the ground work in python that it stopped making sense to do in anything else.
I wanna see if I can't make the functions in Julia instead as the built in linear algebra operations seem very useful