Modularizing functions
Making the functions a proper module.
Additionally, I realized that the reason the sims kept crashing was due to the incompatibility of `Dict{Symbol, V}` and `Dict{String, V}`. The function I was using for setting the parameters was constructing a symbol dict, and then trying to call a string dict, so parameters were undefined.